// src/lib/openapi/adminapi.schema.ts
export interface components {
    schemas: {
// ...
        CreateEntityCorporate: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            entity_type: 'corporate';
            /**
             * Format: email
             * @example random@example.com
             */
            email: string;
            /** @example Blk 447 Clementi Ave 3 #01-201 S() */
            street: string;
            /** @example singapore */
            city: string;
            /** @example 120447 */
            postal_code: string;
            country: components['schemas']['CountryCode2'];
            /** @example Wisoky - Berge */
            company_name: string;
            registration_country: components['schemas']['CountryCode2'];
            incorporation_date: components['schemas']['Date'];
        };
// ...
    };
// ...
}